+++ hazelcast/src/main/java/com/hazelcast/concurrent/lock/client/AbstractLockRequest.java
@@ -39,7 +39,7 @@ public abstract class AbstractLockRequest extends KeyBasedClientRequest
 
     protected Data key;
 
-    private int threadId;
+    private long threadId;
 
     private long ttl = -1;
 
@@ -48,12 +48,12 @@ public abstract class AbstractLockRequest extends KeyBasedClientRequest
     public AbstractLockRequest() {
     }
 
-    public AbstractLockRequest(Data key, int threadId) {
+    public AbstractLockRequest(Data key, long threadId) {
         this.key = key;
         this.threadId = threadId;
     }
 
-    public AbstractLockRequest(Data key, int threadId, long ttl, long timeout) {
+    public AbstractLockRequest(Data key, long threadId, long ttl, long timeout) {
         this.key = key;
         this.threadId = threadId;
         this.ttl = ttl;
@@ -78,7 +78,7 @@ public abstract class AbstractLockRequest extends KeyBasedClientRequest
     }
 
     public void write(PortableWriter writer) throws IOException {
-        writer.writeInt("tid", threadId);
+        writer.writeLong("tid", threadId);
         writer.writeLong("ttl", ttl);
         writer.writeLong("timeout", timeout);
 
@@ -87,7 +87,7 @@ public abstract class AbstractLockRequest extends KeyBasedClientRequest
     }
 
     public void read(PortableReader reader) throws IOException {
-        threadId = reader.readInt("tid");
+        threadId = reader.readLong("tid");
         ttl = reader.readLong("ttl");
         timeout = reader.readLong("timeout");
